Refactor payment request flow
