+++ cardstack/src/screens/PaymentRequest/RequestPaymentConfirmation.tsx
@@ -1,0 +1,220 @@
+import { generateMerchantPaymentUrl } from '@cardstack/cardpay-sdk';
+import React, { useCallback, useMemo, useState } from 'react';
+import CardWalletLogo from '../../assets/cardstackLogo.png';
+import {
+  CopyToast,
+  ToastPositionContainer,
+} from '@rainbow-me/components/toasts';
+import { Icon } from '@rainbow-me/components/icons';
+import {
+  useAccountSettings,
+  useClipboard,
+  useDimensions,
+} from '@rainbow-me/hooks';
+import { Button, Container, Image, Text } from '@cardstack/components';
+import { MerchantInformation } from '@cardstack/types';
+import { shareRequestPaymentLink, splitAddress } from '@cardstack/utils';
+import { ContactAvatar } from '@rainbow-me/components/contacts';
+import { useNavigation } from '@rainbow-me/navigation';
+import { shadow } from '@rainbow-me/styles';
+import Routes from '@rainbow-me/routes';
+import logger from 'logger';
+
+export const RequestPaymentConfirmation = ({
+  amountWithSymbol,
+  address,
+  amountInSpend,
+  merchantInfo,
+}: {
+  amountWithSymbol: string | undefined;
+  address: string;
+  amountInSpend: number;
+  merchantInfo?: MerchantInformation;
+}) => {
+  const [copyCount, setCopyCount] = useState(0);
+
+  const { network } = useAccountSettings();
+  const { navigate } = useNavigation();
+  const { setClipboard } = useClipboard();
+
+  // for now the amount is in spend, but once we get
+  // the currency selector we should pass the amount with currency
+  const paymentRequestLink = useMemo(
+    () =>
+      generateMerchantPaymentUrl({
+        merchantSafeID: address,
+        amount: amountInSpend,
+        network,
+        currency: 'SPD',
+      }),
+    [address, amountInSpend, network]
+  );
+
+  const copyToClipboard = useCallback(() => {
+    setClipboard(paymentRequestLink);
+    setCopyCount(count => count + 1);
+  }, [paymentRequestLink, setClipboard]);
+
+  const handleShareLink = useCallback(async () => {
+    try {
+      await shareRequestPaymentLink(address, paymentRequestLink);
+    } catch (e) {
+      logger.sentry('Payment Request Link share failed', e);
+    }
+  }, [address, paymentRequestLink]);
+
+  // assuming we are using the global native currency,
+  // once it needs to change we probably should receive this as prop
+
+  const showQRCode = useCallback(() => {
+    navigate(Routes.SHOW_QRCODE_MODAL, {
+      value: paymentRequestLink,
+      amountInSpend,
+      amountWithSymbol,
+      name: merchantInfo?.name,
+    });
+  }, [
+    amountInSpend,
+    amountWithSymbol,
+    merchantInfo,
+    navigate,
+    paymentRequestLink,
+  ]);
+
+  return (
+    <>
+      <Container paddingHorizontal={5} width="100%">
+        <Container flexDirection="row" marginTop={8}>
+          <Text color="blueText" fontWeight="bold" paddingTop={1} size="xxs">
+            TO:
+          </Text>
+          <Container flexDirection="row" marginTop={4}>
+            <Container paddingHorizontal={1} paddingTop={4}>
+              <ContactAvatar
+                color={merchantInfo?.color}
+                size="smaller"
+                textColor={merchantInfo?.textColor}
+                value={merchantInfo?.name}
+              />
+            </Container>
+            <Container paddingLeft={2}>
+              <Text color="blueText" size="smallest" textTransform="uppercase">
+                Merchant
+              </Text>
+              <Text fontSize={15} fontWeight="bold">
+                {merchantInfo?.name}
+              </Text>
+              <Text
+                color="blueText"
+                fontFamily="RobotoMono-Regular"
+                marginTop={2}
+                size="small"
+              >
+                {splitAddress(address).twoLinesAddress}
+              </Text>
+            </Container>
+          </Container>
+        </Container>
+        <Container
+          backgroundColor="grayCardBackground"
+          borderRadius={10}
+          marginTop={8}
+          paddingVertical={5}
+        >
+          <Container>
+            <Container
+              alignItems="center"
+              alignSelf="center"
+              flexDirection="row"
+              paddingLeft={4}
+            >
+              <Icon name="qrCodeBig" />
+              <Text
+                fontSize={15}
+                fontWeight="600"
+                letterSpacing={0.15}
+                lineHeight={20}
+                paddingLeft={3}
+              >
+                Let your customer scan a {'\n'}QR code to pay
+              </Text>
+            </Container>
+          </Container>
+          <Container marginTop={6} paddingHorizontal={5}>
+            <Button onPress={showQRCode} width="100%">
+              Show QR code
+            </Button>
+            <Container
+              alignItems="center"
+              alignSelf="center"
+              flexDirection="row"
+              marginTop={7}
+            >
+              <Icon name="link" />
+              <Text
+                fontSize={15}
+                fontWeight="600"
+                letterSpacing={0.15}
+                lineHeight={20}
+                paddingLeft={3}
+              >
+                Or send your customer {'\n'}the link to pay
+              </Text>
+            </Container>
+          </Container>
+          <Container
+            flex={1}
+            flexDirection="row"
+            flexWrap="wrap"
+            marginTop={6}
+            paddingHorizontal={5}
+            width="100%"
+          >
+            <Container flex={1} paddingRight={2}>
+              <Button onPress={copyToClipboard} variant="small" width="100%">
+                Copy Link
+              </Button>
+            </Container>
+            <Container flex={1} paddingLeft={2}>
+              <Button onPress={handleShareLink} variant="small" width="100%">
+                Share Link
+              </Button>
+            </Container>
+          </Container>
+        </Container>
+      </Container>
+      {copyCount > 0 ? (
+        <ToastPositionContainer>
+          <CopyToast copiedText="Payment Request Link" copyCount={copyCount} />
+        </ToastPositionContainer>
+      ) : null}
+    </>
+  );
+};
+
+export const RequestPaymentConfirmationFooter = () => {
+  const { isTallPhone } = useDimensions();
+
+  return (
+    <Container
+      alignSelf="flex-end"
+      backgroundColor="white"
+      bottom={isTallPhone ? 70 : 0}
+      paddingBottom={5}
+      paddingTop={3}
+      style={shadow.buildAsObject(0, -1, 2, 'rgba(0, 0, 0, 0.25)', 1)}
+      width="100%"
+    >
+      <Container alignItems="center" justifyContent="center">
+        <Image height={30} source={CardWalletLogo} width={30} />
+        <Text marginTop={1} size="xs" textAlign="center">
+          Your customer must have the {'\n'}
+          <Text fontWeight="bold" size="xs">
+            Card Wallet mobile app
+          </Text>{' '}
+          installed.
+        </Text>
+      </Container>
+    </Container>
+  );
+};
