+++ cardstack/src/screens/PaymentRequest/index.ts
@@ -1,0 +1,1 @@
+export { default as PaymentRequestExpandedState } from './PaymentRequestExpandedState';
