+++ cardstack/src/screens/index.ts
@@ -8,3 +8,4 @@ export { default as SendSheetDepot } from './SendSheetDepot/SendSheetDepot';
 export { default as ShowQRCodeModal } from './ShowQRCodeModal';
 export { default as PayMerchant } from './PayMerchant/PayMerchant';
 export { default as CurrencySelectionGlobalModal } from './CurrencySelectionGlobalModal';
+export { default as PaymentRequestExpandedState } from './PaymentRequest/PaymentRequestExpandedState';
