+++ src/components/expanded-state/index.js
@@ -10,5 +10,4 @@ export { default as UnclaimedRevenueExpandedState } from './UnclaimedRevenueExpa
 export { default as AvailableBalancesExpandedState } from './AvailableBalancesExpandedState';
 export { default as MerchantTransactionExpandedState } from './MerchantTransactionExpandedState';
 export { default as PaymentConfirmationExpandedState } from './PaymentConfirmationExpandedState';
-export { default as PaymentRequestExpandedState } from './PaymentRequestExpandedState';
 export { default as RecentActivityExpandedState } from './RecentActivityExpandedState';
